requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
psycopg2-binary
python-dotenv
requests